Moved the Planning Interval components into a components folder under the Planning Interval folder.
