+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/[key]/objectives/[objectiveKey]/page.tsx
@@ -6,7 +6,7 @@ import { useMemo, useState } from 'react'
 import PlanningIntervalObjectiveDetails from './planning-interval-objective-details'
 import { useDocumentTitle } from '@/src/app/hooks/use-document-title'
 import useAuth from '@/src/app/components/contexts/auth'
-import EditPlanningIntervalObjectiveForm from '../../../../components/edit-planning-interval-objective-form'
+import EditPlanningIntervalObjectiveForm from '../../../components/edit-planning-interval-objective-form'
 import { ItemType } from 'antd/es/menu/interface'
 import DeletePlanningIntervalObjectiveForm from './delete-planning-interval-objective-form'
 import { authorizePage } from '@/src/app/components/hoc'
