+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/[key]/plan-review/team-plan-review.tsx
@@ -14,7 +14,7 @@ import { SegmentedLabeledOption } from 'antd/es/segmented'
 import {
   PlanningIntervalObjectiveDetailsDrawer,
   PlanningIntervalObjectivesTimeline,
-} from '../../../components'
+} from '../../components'
 import {
   useGetPlanningIntervalCalendar,
   useGetPlanningIntervalRisksByTeamId,
