+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/components/edit-planning-interval-form.tsx
@@ -2,7 +2,7 @@
 
 import { Form, Input, Modal, Switch, message } from 'antd'
 import { useCallback, useEffect, useState } from 'react'
-import useAuth from '../../components/contexts/auth'
+import useAuth from '../../../components/contexts/auth'
 import {
   PlanningIntervalDetailsDto,
   UpdatePlanningIntervalRequest,
