+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/components/edit-planning-interval-objective-form.tsx
@@ -13,7 +13,7 @@ import {
   message,
 } from 'antd'
 import { useCallback, useEffect, useState } from 'react'
-import useAuth from '../../components/contexts/auth'
+import useAuth from '../../../components/contexts/auth'
 import {
   PlanningIntervalObjectiveDetailsDto,
   UpdatePlanningIntervalObjectiveRequest,
