+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/page.tsx
@@ -6,7 +6,7 @@ import Link from 'next/link'
 import { useCallback, useMemo, useState } from 'react'
 import { useDocumentTitle } from '../../hooks/use-document-title'
 import dayjs from 'dayjs'
-import { CreatePlanningIntervalForm } from '../components'
+import { CreatePlanningIntervalForm } from './components'
 import useAuth from '../../components/contexts/auth'
 import { Button } from 'antd'
 import { useGetPlanningIntervals } from '@/src/services/queries/planning-queries'
